rudimentary solution
Note that the program relies on the fact that syscall `write` and `read`
to pipes always success, in the sense that they write configured bytes
or read all the currently available bytes, resp.

That being said I suspect much of the time is indeed in user space, due
to the string processing.